ci: image builder support (#21)
* ci: add support for additional ARM64 arch strings

Add armv8* (such as armv8l) support to TARGETARCH emulation.  The pi-gen project reports this arch
when building RPi images.  Most likely cause is the builder is running in 32bit mode
(https://android.stackexchange.com/a/217477).

* ci: add additional support for building in different environments

- Refactored common scripts to 'script/_common'
- Add missing dependencies for building on AWS Ubuntu images.
- Add support for "docker-less" elevated privileges for sections that do
not require docker.  Useful for pre-building some components on an RPi
image for example.

* docs: remove container reference from BLE tshooting page